refactor: 优化宏模块结构
- 删除未使用的 application-macro crate
- 将数据库查询宏移到单独的 macros.rs 模块
- 使用 #[macro_export] + use crate::{...} 替代 pub(crate) use
- 符合 Rust 最佳实践，移除 #[allow] 属性
