--- application-rs/application-macro/.gitignore
@@ -1,19 +1,0 @@
-# Windows
-[Dd]esktop.ini
-Thumbs.db
-$RECYCLE.BIN/
-
-# macOS
-.DS_Store
-.fseventsd
-.Spotlight-V100
-.TemporaryItems
-.Trashes
-
-# Added by cargo
-
-/target
-
-# 开发
-config.toml
-http-client.private.env.json
