--- application-rs/application-macro/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "application-macro"
-rust-version.workspace = true
-version.workspace = true
-edition.workspace = true
-publish.workspace = true
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "~2.0.0", features = ["full"] }
-quote = { version = "~1.0.0" }
-proc-macro2 = { version = "~1.0.0" }
